refactor: refactor Electron main process 1. Refactor Electron main process into SRP modules. 2. Migrate nx to latest version. 3. Add loading screen before loading into main screen.
