+++ main-process/constants.cjs
@@ -1,0 +1,9 @@
+'use strict';
+
+const URLs = ['http://localhost:5000', 'http://localhost:3000'];
+const ROOT_DATABASE_NAME = 'database.sqlite3';
+
+module.exports = {
+  URLs,
+  ROOT_DATABASE_NAME,
+};
