+++ main-process/environment-utils.cjs
@@ -1,0 +1,6 @@
+'use strict';
+function getEnvironment() {
+  return process.env.NODE_ENV || 'prod';
+}
+
+module.exports = { getEnvironment };
