+++ main-process/file-utils.cjs
@@ -1,0 +1,10 @@
+'use strict';
+const { writeFileSync } = require('fs');
+
+function writePath(filePath, content) {
+  writeFileSync(filePath, content, 'utf8');
+}
+
+module.exports = {
+  writePath,
+};
