--- nest-backend/webpack.config.js
@@ -1,8 +1,0 @@
-const { composePlugins, withNx } = require('@nx/webpack');
-
-// Nx plugins for webpack.
-module.exports = composePlugins(withNx(), (config) => {
-  // Update the webpack config as needed here.
-  // e.g. `config.plugins.push(new MyPlugin())`
-  return config;
-});
