+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "elec-ng-nest",
   "version": "0.0.0",
-  "main": "main.js",
+  "main": "main.cjs",
   "license": "MIT",
   "type": "module",
   "author": "Guan Xin Wang",
@@ -15,77 +15,76 @@
     "build": "npm run build-front && npm run build-back",
     "build-prod-front": "nx build ng-tracker --configuration production --optimization --base-href ./",
     "build-prod-back": "cross-env-shell NODE_ENV=prod PORT=5000 \"nx build nest-backend\"",
-    "build-prod": "npm run build-prod-front && npm run build-prod-back && node hooks.js",
+    "build-prod": "npm run build-prod-front && npm run build-prod-back && node hooks.cjs",
     "make": "cross-env-shell NODE_ENV=prod PORT=5000 \"npm run build-prod && electron-forge make\"",
     "package": "cross-env-shell NODE_ENV=prod \"npm run build-prod && electron-forge package\""
   },
   "private": true,
   "devDependencies": {
-    "@analogjs/vite-plugin-angular": "^1.7.3",
-    "@angular-devkit/build-angular": "~17.3.0",
-    "@angular-devkit/core": "~17.3.0",
-    "@angular-devkit/schematics": "~17.3.0",
-    "@angular-eslint/eslint-plugin": "~17.3.0",
-    "@angular-eslint/eslint-plugin-template": "~17.3.0",
-    "@angular-eslint/template-parser": "~17.3.0",
-    "@angular/animations": "~17.3.0",
-    "@angular/cli": "~17.3.0",
-    "@angular/common": "~17.3.0",
-    "@angular/compiler": "~17.3.0",
-    "@angular/compiler-cli": "~17.3.0",
-    "@angular/core": "~17.3.0",
-    "@angular/forms": "~17.3.0",
-    "@angular/language-service": "~17.3.0",
-    "@angular/platform-browser": "~17.3.0",
-    "@angular/platform-browser-dynamic": "~17.3.0",
-    "@angular/router": "~17.3.0",
+    "@angular-devkit/build-angular": "18.2.4",
+    "@angular-devkit/core": "18.2.4",
+    "@angular-devkit/schematics": "18.2.4",
+    "@angular-eslint/eslint-plugin": "18.0.1",
+    "@angular-eslint/eslint-plugin-template": "18.0.1",
+    "@angular-eslint/template-parser": "18.0.1",
+    "@angular/animations": "18.2.4",
+    "@angular/cli": "18.2.3",
+    "@angular/common": "18.2.4",
+    "@angular/compiler": "18.2.4",
+    "@angular/compiler-cli": "18.2.4",
+    "@angular/core": "18.2.4",
+    "@angular/forms": "18.2.4",
+    "@angular/language-service": "18.2.4",
+    "@angular/platform-browser": "18.2.4",
+    "@angular/platform-browser-dynamic": "18.2.4",
+    "@angular/router": "18.2.4",
     "@electron-forge/cli": "^7.4.0",
     "@electron-forge/maker-deb": "^7.4.0",
     "@electron-forge/maker-rpm": "^7.4.0",
     "@electron-forge/maker-squirrel": "^7.4.0",
     "@electron-forge/maker-zip": "^7.4.0",
     "@electron-forge/plugin-auto-unpack-natives": "^7.4.0",
-    "@fortawesome/angular-fontawesome": "^0.14.1",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
     "@fortawesome/free-solid-svg-icons": "^6.5.2",
     "@nestjs/schematics": "^10.0.1",
     "@nestjs/testing": "^10.0.2",
-    "@nx/angular": "^19.0.3",
-    "@nx/eslint": "19.0.3",
-    "@nx/eslint-plugin": "19.0.3",
-    "@nx/jest": "19.0.3",
-    "@nx/js": "19.0.3",
-    "@nx/nest": "^19.0.3",
-    "@nx/node": "19.0.3",
-    "@nx/vite": "^19.0.3",
-    "@nx/web": "19.0.3",
-    "@nx/webpack": "19.0.3",
-    "@nx/workspace": "19.0.3",
-    "@schematics/angular": "~17.3.0",
-    "@swc-node/register": "~1.8.0",
-    "@swc/core": "~1.3.85",
-    "@swc/helpers": "~0.5.2",
-    "@types/jest": "^29.4.0",
+    "@nx/angular": "19.7.2",
+    "@nx/eslint": "19.7.2",
+    "@nx/eslint-plugin": "19.7.2",
+    "@nx/jest": "19.7.2",
+    "@nx/js": "19.7.2",
+    "@nx/nest": "19.7.2",
+    "@nx/node": "19.7.2",
+    "@nx/vite": "19.7.2",
+    "@nx/web": "19.7.2",
+    "@nx/webpack": "19.7.2",
+    "@nx/workspace": "19.7.2",
+    "@schematics/angular": "18.2.4",
+    "@swc-node/register": "1.9.2",
+    "@swc/core": "1.5.7",
+    "@swc/helpers": "0.5.13",
+    "@types/jest": "29.5.12",
     "@types/node": "18.16.9",
     "@types/uuid": "^9.0.8",
-    "@typescript-eslint/eslint-plugin": "^7.3.0",
-    "@typescript-eslint/parser": "^7.3.0",
+    "@typescript-eslint/eslint-plugin": "7.16.0",
+    "@typescript-eslint/parser": "7.16.0",
     "@vitest/coverage-v8": "^1.0.4",
     "@vitest/ui": "^1.3.1",
     "cross-env": "^7.0.3",
-    "electron": "^30.0.8",
+    "electron": "^32.0.2",
     "electron-is-dev": "^3.0.1",
     "eslint": "~8.57.0",
     "eslint-config-prettier": "^9.0.0",
-    "jest": "^29.4.1",
-    "jest-environment-jsdom": "^29.4.1",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
     "jest-environment-node": "^29.4.1",
-    "jest-preset-angular": "~14.0.3",
-    "nx": "19.0.3",
+    "jest-preset-angular": "14.1.1",
+    "nx": "19.7.2",
     "prettier": "^2.6.2",
     "sqlite3": "^5.1.7",
     "ts-jest": "^29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "~5.4.2",
+    "typescript": "5.5.4",
     "vite": "~5.0.0",
     "vitest": "^1.3.1",
     "webpack-cli": "^5.1.4",
